Backend/requirements.txt
# Career Roadmap AI Pipeline Dependencies

# Core dependencies
pydantic>=2.0.0,<3.0.0
openai>=1.0.0,<2.0.0
pathlib2>=2.3.0  # For Python < 3.4 compatibility

# Optional dependencies for enhanced functionality
python-dotenv>=1.0.0  # For environment variable management
requests>=2.31.0  # For HTTP requests if needed
jsonschema>=4.17.0  # For additional JSON schema validation

# Development dependencies (optional)
pytest>=7.4.0  # For testing
black>=23.0.0  # For code formatting
flake8>=6.0.0  # For linting
mypy>=1.5.0  # For type checking

# Logging and monitoring (optional)
structlog>=23.1.0  # For structured logging